[CHG] Shop and Item management

+++ trunk/ProductTracker/ProductTracker.Web/ManageShops.aspx.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Web;
 
 #endregion
 
@@ -21,12 +22,12 @@ namespace ProductTracker.Web
 
         private void ClearFields()
         {
-            TextBoxBodyShopsId.Text = "";
             TextBoxBodyShopsName.Text = "";
             TextBoxBodyShopsOwner.Text = "";
             TextBoxBodyShopsAddress.Text = "";
             TextBoxBodyShopsCity.Text = "";
             TextBoxBodyShopsPostalCode.Text = "";
+            EnableAdd();
         }
 
         private void PopulateShops()
@@ -36,6 +37,20 @@ namespace ProductTracker.Web
             DropDownListShopList.DataSource = shops;
             DropDownListShopList.DataBind();
             ClearFields();
+        }
+
+        private void EnableAdd()
+        {
+            LinkButtonBodyShopsSubmit.Visible = true;
+            LinkButtonBodyShopsDelete.Visible = false;
+            LinkButtonBodyShopsUpdate.Visible = false;
+        }
+
+        private void EnableUpdate()
+        {
+            LinkButtonBodyShopsSubmit.Visible = false;
+            LinkButtonBodyShopsDelete.Visible = true;
+            LinkButtonBodyShopsUpdate.Visible = true;
         }
 
         private void SetElementValues()
@@ -48,7 +63,6 @@ namespace ProductTracker.Web
                 {
                     List<Setting> setting = page.Setting;
                     HyperLinkMain.Text = settingsManager.GetSettingValue("HyperLinkMain", setting);
-                    LabelBodyShopsId.Text = settingsManager.GetSettingValue("LabelBodyShopsId", setting);
                     LabelBodyShopsName.Text = settingsManager.GetSettingValue("LabelBodyShopsName", setting);
                     LabelBodyShopsAddress.Text = settingsManager.GetSettingValue("LabelBodyShopsAddress", setting);
                     LabelBodyShopsOwner.Text = settingsManager.GetSettingValue("LabelBodyShopsOwner", setting);
@@ -57,8 +71,12 @@ namespace ProductTracker.Web
                     LabelBodyShopsIsCompany.Text = settingsManager.GetSettingValue("LabelBodyShopsIsCompany", setting);
                     LabelShopList.Text = settingsManager.GetSettingValue("LabelShopList", setting);
                     LinkButtonShopList.Text = settingsManager.GetSettingValue("LinkButtonShopList", setting);
-                    LinkButtonBodyShopsSubmit.Text = settingsManager.GetSettingValue("LinkButtonBodyShopsSubmit", setting);
-                    LinkButtonBodyShopsDelete.Text = settingsManager.GetSettingValue("LinkButtonBodyShopsDelete", setting);
+                    LinkButtonBodyShopsSubmit.Text = settingsManager.GetSettingValue("LinkButtonBodyShopsSubmit",
+                                                                                     setting);
+                    LinkButtonBodyShopsDelete.Text = settingsManager.GetSettingValue("LinkButtonBodyShopsDelete",
+                                                                                     setting);
+                    LinkButtonBodyShopsUpdate.Text = settingsManager.GetSettingValue("LinkButtonBodyShopsUpdate",
+                                                                                     setting);
                     LinkButtonAdd.Text = settingsManager.GetSettingValue("LinkButtonAdd", setting);
                 }
             }
@@ -68,15 +86,23 @@ namespace ProductTracker.Web
         {
             string selectedValue = DropDownListShopList.SelectedValue;
             DataBase dataBase = new DataBase();
-            Shop shop = dataBase.GetShop(selectedValue);
+            Shop shop = dataBase.GetShopByName(selectedValue);
             if (shop != null)
             {
-                TextBoxBodyShopsId.Text = shop.Id.ToString();
                 TextBoxBodyShopsName.Text = shop.Name;
                 TextBoxBodyShopsOwner.Text = shop.Owner;
                 TextBoxBodyShopsAddress.Text = shop.Address;
                 TextBoxBodyShopsCity.Text = shop.City;
                 TextBoxBodyShopsPostalCode.Text = shop.PostalCode.ToString();
+                HttpCookie cookie = Request.Cookies[CookieName];
+                if (cookie == null)
+                {
+                    cookie = new HttpCookie(CookieName);
+                }
+                cookie["Name"] = shop.Id.ToString();
+                cookie.Expires = DateTime.Now.AddMinutes(15);
+                Response.Cookies.Add(cookie);
+                EnableUpdate();
             }
             else
             {
@@ -88,33 +114,54 @@ namespace ProductTracker.Web
         {
             DataBase dataBase = new DataBase();
             Shop shop = new Shop
-            {
-                Id = Guid.NewGuid(),
-                Name = TextBoxBodyShopsName.Text.Trim(),
-                Address = TextBoxBodyShopsAddress.Text.Trim(),
-                City = TextBoxBodyShopsCity.Text.Trim(),
-                Owner = TextBoxBodyShopsOwner.Text.Trim(),
-                PostalCode = Int32.Parse(TextBoxBodyShopsPostalCode.Text.Trim()),
-                IsCompany = CheckBoxBodyShopsIsCompany.Checked,
-            };
-            if (TextBoxBodyShopsId.Text.Length > 8)
-            {
-                shop.Id = new Guid(TextBoxBodyShopsId.ToString());
-                dataBase.UpdateShop(shop);
-            }
-            else
-            {
-                dataBase.InsertShop(shop);
-            }
+                {
+                    Id = Guid.NewGuid(),
+                    Name = TextBoxBodyShopsName.Text.Trim(),
+                    Address = TextBoxBodyShopsAddress.Text.Trim(),
+                    City = TextBoxBodyShopsCity.Text.Trim(),
+                    Owner = TextBoxBodyShopsOwner.Text.Trim(),
+                    PostalCode = Int32.Parse(TextBoxBodyShopsPostalCode.Text.Trim()),
+                    IsCompany = CheckBoxBodyShopsIsCompany.Checked,
+                };
+            dataBase.InsertShop(shop);
             PopulateShops();
+            ClearFields();
         }
 
         protected void LinkButtonBodyShopsDelete_Click(object sender, EventArgs e)
         {
+            HttpCookie cookie = Request.Cookies[CookieName];
+            if (cookie == null)
+            {
+                return;
+            }
             DataBase dataBase = new DataBase();
-            string id = TextBoxBodyShopsId.Text;
+            string id = cookie["Name"];
             dataBase.DeleteShop(id);
             PopulateShops();
+        }
+
+        protected void LinkButtonBodyShopsUpdate_Click(object sender, EventArgs e)
+        {
+            HttpCookie cookie = Request.Cookies[CookieName];
+            if (cookie == null)
+            {
+                return;
+            }
+            DataBase dataBase = new DataBase();
+            Shop shop = new Shop
+                {
+                    Name = TextBoxBodyShopsName.Text.Trim(),
+                    Address = TextBoxBodyShopsAddress.Text.Trim(),
+                    City = TextBoxBodyShopsCity.Text.Trim(),
+                    Owner = TextBoxBodyShopsOwner.Text.Trim(),
+                    PostalCode = Int32.Parse(TextBoxBodyShopsPostalCode.Text.Trim()),
+                    IsCompany = CheckBoxBodyShopsIsCompany.Checked,
+                    Id = new Guid(cookie["Name"]),
+                };
+            dataBase.UpdateShop(shop);
+            PopulateShops();
+            ClearFields();
         }
 
         protected void LinkButtonAdd_Click(object sender, EventArgs e)
@@ -123,5 +170,7 @@ namespace ProductTracker.Web
         }
 
         //private string statusShopNotFound = "Shop not found!";
+
+        private const string CookieName = "CookieShopId";
     }
 }
